requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
boto3==1.26.68
botocore==1.29.68
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
Django==4.1.5
django-cors-headers==3.13.0
django-scheduler==0.10.0
django-storages==1.13.2
frozenlist==1.3.3
future==0.18.3
icalendar==5.0.4
idna==3.4
jmespath==1.0.1
line-bot-sdk==2.4.1
multidict==6.0.4
mysql-connector==2.2.9
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
yarl==1.8.2
